Fixed issue #4932 : unable to delete single response Dev : slect the file to delete only if there are one file or more
git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey@9745 b72ed6b6-b9f8-46b5-92b4-906544132732

+++ admin/browse.php
@@ -366,20 +366,22 @@ elseif ($subaction == "all")
             if ($field['type'] == "|" && strpos($field['fieldname'], "_filecount") == 0)
                 $fuqtquestions[] = $field['fieldname'];
         }
-
-        // find all responses (filenames) to the fuqt questions
-        $query="SELECT " . implode(", ", $fuqtquestions) . " FROM $surveytable where id={$_POST['deleteanswer']}";
-        $responses = db_execute_assoc($query) or safe_die("Could not fetch responses<br />$query<br />".$connect->ErrorMsg());
-
-        while($json = $responses->FetchRow())
+        if (count($fuqtquestions)>0)
         {
-            foreach ($fuqtquestions as $fieldname)
+            // find all responses (filenames) to the fuqt questions
+            $query="SELECT " . implode(", ", $fuqtquestions) . " FROM $surveytable where id={$_POST['deleteanswer']}";
+            $responses = db_execute_assoc($query) or safe_die("Could not fetch responses<br />$query<br />".$connect->ErrorMsg());
+
+            while($json = $responses->FetchRow())
             {
-                $phparray = json_decode($json[$fieldname]);
-                foreach($phparray as $metadata)
+                foreach ($fuqtquestions as $fieldname)
                 {
-                    $path = dirname(getcwd())."/upload/surveys/".$surveyid."/files/";
-                    unlink($path.$metadata->filename); // delete the file
+                    $phparray = json_decode($json[$fieldname]);
+                    foreach($phparray as $metadata)
+                    {
+                        $path = dirname(getcwd())."/upload/surveys/".$surveyid."/files/";
+                        unlink($path.$metadata->filename); // delete the file
+                    }
                 }
             }
         }
